+++ packages/widget/src/pages/SwapPage/index.ts
@@ -1,1 +1,2 @@
 export * from './SwapPage';
+export * from './utils';
